remove togle and add top 3 Frontliner
