+++ package.json
@@ -48,6 +48,7 @@
     "react-no-ssr": "^1.1.0",
     "react-router-dom": "^6.4.0",
     "react-toastify": "^11.0.5",
+    "recharts": "^2.15.3",
     "tailwindcss-rtl": "^0.9.0",
     "typescript": "^4.9.4",
     "web-vitals": "^2.1.4"
